refactor: extract colorWithOpacity to shared app/util/colors utility
Move colorWithOpacity from Predictions/utils/ to app/util/colors/
and replace the duplicated copy in PredictSportTeamGradient.tsx with
the shared import. Eliminates drift risk from maintaining two
identical 85-line functions.
